Move collect_fields function
